+++ src/core/object/watch/engines/proxy.ts
@@ -366,6 +366,12 @@ export function set(obj: object, path: WatchPath, value: unknown, handlers: Watc
 		prop = normalizedPath[normalizedPath.length - 1],
 		refPath = normalizedPath.slice(0, -1);
 
+	if (normalizedPath.length > 1 && Object.get(obj, refPath) == null) {
+		unwrappedObj[muteLabel] = true;
+		Object.set(obj, refPath, {});
+		unwrappedObj[muteLabel] = false;
+	}
+
 	const ref = Object.get(
 		getOrCreateLabelValueByHandlers(unwrappedObj, toProxyObject, handlers) ?? unwrappedObj,
 		refPath
@@ -376,7 +382,6 @@ export function set(obj: object, path: WatchPath, value: unknown, handlers: Watc
 		type = getProxyType(ref);
 
 	switch (type) {
-		case null:
 		case 'set':
 			throw new TypeError('Invalid data type to watch');
 
@@ -433,7 +438,7 @@ export function unset(obj: object, path: WatchPath, handlers: WatchHandlersSet):
 
 	switch (type) {
 		case null:
-			throw new TypeError('Invalid data type to watch');
+			return;
 
 		case 'array':
 			(<unknown[]>ref).splice(Number(prop), 1);